+++ src/main/java/ait/cohort34/smartyflip/card/dao/CardRepository.java
@@ -11,6 +11,9 @@ package ait.cohort34.smartyflip.card.dao;
 import ait.cohort34.smartyflip.card.model.Card;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CardRepository extends JpaRepository<Card, Long> {
+/**
+ * This interface represents a repository for managing Card objects.
+ */
+public interface CardRepository extends JpaRepository<Card, Integer> {
 
 }
